requirements.txt
transformers
sentencepiece
python-dotenv
